Move Health check binding to new HealthService.
